backend/requirements.txt
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.16.1
google-auth==2.49.1
python-dotenv==1.2.2
requests==2.32.5
psycopg2-binary

# LangChain ecosystem (compatible with langgraph)
langgraph==0.0.69
langchain==0.2.1
langchain-core==0.2.33
langchain-community==0.2.1
langchain-openai==0.1.8

# Google GenAI (UPDATED to match langchain-google-genai)
langchain-google-genai==1.0.10
google-generativeai==0.7.2
google-ai-generativelanguage==0.6.6

python-dateutil==2.8.2